Simplify code.
