+++ util/gen-types/src/core.rs
@@ -2,11 +2,6 @@
 
 #![allow(clippy::from_over_into)]
 
-#[cfg(feature = "std")]
-mod std_env;
-#[cfg(feature = "std")]
-pub use std_env::*;
-
 use crate::packed;
 
 /// Specifies how the script `code_hash` is used to match the script code and how to run the code.
@@ -56,5 +51,33 @@ impl Into<packed::Byte> for ScriptHashType {
     #[inline]
     fn into(self) -> packed::Byte {
         Into::<u8>::into(self).into()
+    }
+}
+
+#[cfg(feature = "std")]
+mod std_mod {
+    use crate::{core::ScriptHashType, packed};
+    use ckb_error::OtherError;
+
+    impl TryFrom<u8> for ScriptHashType {
+        type Error = OtherError;
+
+        fn try_from(v: u8) -> Result<Self, Self::Error> {
+            match v {
+                0 => Ok(ScriptHashType::Data),
+                1 => Ok(ScriptHashType::Type),
+                2 => Ok(ScriptHashType::Data1),
+                4 => Ok(ScriptHashType::Data2),
+                _ => Err(OtherError::new(format!("Invalid script hash type {v}"))),
+            }
+        }
+    }
+
+    impl TryFrom<packed::Byte> for ScriptHashType {
+        type Error = OtherError;
+
+        fn try_from(v: packed::Byte) -> Result<Self, Self::Error> {
+            Into::<u8>::into(v).try_into()
+        }
     }
 }
